fix(tests): Fixed attribute test-suite organization by moving and isolating core attribute coverage into dedicated test classes for clearer maintenance.
